Feat(Pet-Engine): Added persistence for the pet stats and equipments (Firestore implementation).

+++ app/src/main/java/com/android/sample/screens/CreatureEnvironment.kt
@@ -171,7 +171,7 @@ fun CreatureStatsCard(stats: CreatureStats, modifier: Modifier = Modifier) {
 @Composable
 private fun StatRow(
     title: String,
-    value: Int,
+    value: Float,
     icon: ImageVector,
     barColor: Color,
 ) {
